requirements.txt
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
grep-ast==0.3.2
networkx==3.2.1
pygments==2.18.0
tqdm
datasets
openai==1.42.0
tiktoken==0.7.0
libcst==1.4.0
pyvis
langchain
langgraph
langchain_openai
python-dotenv
scikit-learn
sentence-transformers
spacy
scipy
python-louvain
numpy
jinja2
markdown
jsonschema
python-magic
weasyprint
# python -m spacy download en_core_web_md